fix(console):  1. 完成 cluster 主要三个类的 mapper  单元测试 2. 完成 cluster runtime 等 关联穷举 实现：BuildFullSceneData

+++ eventmesh-dashboard-console/src/main/java/org/apache/eventmesh/dashboard/console/entity/cases/ResourcesConfigEntity.java
@@ -26,13 +26,13 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = true)
 public class ResourcesConfigEntity extends CaseEntity {
 
-    private Integer cpuNum;
+    private Float cpuNum;
 
-    private Integer memNum;
+    private Float memNum;
 
-    private Integer diskNum;
+    private Float diskNum;
 
-    private Integer gpuNum;
+    private Float gpuNum;
 
 
 }
